requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
Pillow==4.1.1
django-jet==1.0.6
pytz==2017.2
olefile==0.44
djangorestframework==3.6.3
django-cors-headers==2.1.0